Work on Shiphealth, Safezone, Selfdestruct
Modified all of them to bertter support maps like rpg_runabout with more than one ship
both selfdestruct and shiphealth now allow for a safezones targetname to be added that is freed before called to list
that required modifications to the associated commands.
Now selfdestruct start asks for a safezone as it's 2nd to last argument and requires it if any safezone is spawnflaged as SHIP
shipdamage now has a 2nd argument that specifies the healthEnt it shall deal to. It requires that argument if it finds at least 2 healthEnts
shiphealth now displays the status of all healthEnts that exist and displays it's targetname in line one of the output
There ar also a few other things associated with this (e.g. healthEnt now frees at spawn if it misses required parameters)

Expanded description of target_safezone

Signed-off-by: Harry Young <hendrik.gerritzen@googlemail.com>