More work on selfdestruct
Added a new command selfdestructcountdown duration [safezone] [target]
It'll do exactly the same as selfdestruct start with the exception that it'll force the entity to think every 0.1 secs. to not do this accidently It got it's own command.
It uses the cp-command for display and disables admin_centerprint autoamtically.

On a more general selfdestruct note I got rid of those long warning sentences in favour of ^1Self Destruct in %.0f:%2.0f

Signed-off-by: Harry Young <hendrik.gerritzen@googlemail.com>