Definitions and borg2-elevator
Updated more definitions
Started to work on an entity that manages the 2-part-elevator
Spawnfunc on that works fine (including failsafes)
but thinkfunc buggs out on me (can't say I'm surprised, first time I'm working on trajectories in this way)
I think the problem lies in transfering the hardcoded vectorized speed to the entity.

spawn-nextthink is set to -1 for now.

Signed-off-by: Harry Young <hendrik.gerritzen@googlemail.com>