Started to add HUD-Element for Selfdestruct, however...
there seems to be an issue with the way I selected to transfer the countdown duration to the client (via event)
In server side Execution I run by the g_addevent just fine, however on the client side the event is not called.

there is also a new definition for trigger_airlock.

Signed-off-by: Harry Young <hendrik.gerritzen@googlemail.com>